Rename authentication to signature
